Update index.md
Just adding in my 2 cents here, the MoFi4500 with these two LDPAs are a great combo for rural internet usage. Some extra information regarding MIMO that may be helpful for other rural_internet folks trying to understand the technology, not sure where this would best be added but wanted to at least give you this information if it ends up being helpful at all. 

What causes a router to get low data speeds is essentially the receiving antennas aren't terribly efficient. They pick up a fair amount of "Noise" along with their target signal. 
So really the only thing your router needs to perform better, is a better, cleaner signal coming in. Think of noise signal as taking up a percentage of your bandwidth, if you have a high noise ratio, that’s a low amount of signal in amongst the noise, resulting in a severely decreased bandwidth. 

This is why adding a booster will actually reduce your peak download speeds by about 30%. Even though it's providing a stronger and more stable signal, purely because it adds a significant noise factor.

However, using multiple high powered outdoor antennas means you'll get, firstly, much more signal into each antenna compared to what the little built-in ones can do inside the house. Secondly you're drastically reducing the noise ratio on the signal, because of the directionality of the antennas and the fact that they're cross polarized (so less other signal leaks in). Finally, by feeding both antenna ports, you're making sure you get a complete multi-channel connection (this is the MIMO).  

Of course it's difficult to promise any fixed amount of performance and one shouldn't expect miracles, but this will definitely get you the best performance that is possible, both in terms of latency and packet loss, as well as actual download and upload speeds.